add getIntegrationGuideMarkdown
