+++ apps/web/app/(ee)/app.dub.co/(new-program)/[slug]/program/new/connect/page.tsx
@@ -1,9 +1,8 @@
+import { getIntegrationGuideMarkdown } from "@/lib/get-integration-guide-markdown";
 import { Guide } from "@/ui/guides/guide";
 import { GuideList } from "@/ui/guides/guide-list";
 import { guides, IntegrationGuide } from "@/ui/guides/integrations";
-import { readFileSync } from "fs";
 import { notFound } from "next/navigation";
-import { join } from "path";
 import { Suspense } from "react";
 import { StepPage } from "../step-page";
 
@@ -11,7 +10,7 @@ export default async function Page({
   searchParams,
 }: {
   searchParams: {
-    guide: string;
+    guide?: string;
   };
 }) {
   let { guide } = searchParams;
@@ -25,7 +24,7 @@ export default async function Page({
       notFound();
     }
 
-    markdown = getMarkdown(selectedGuide.key);
+    markdown = await getIntegrationGuideMarkdown(selectedGuide.key);
 
     if (!markdown) {
       notFound();
@@ -49,19 +48,4 @@ export default async function Page({
       </Suspense>
     </StepPage>
   );
-}
-
-function getMarkdown(guideKey: string): string | null {
-  const markdownPath = join(
-    process.cwd(),
-    "integration-guides",
-    `${guideKey}.md`,
-  );
-
-  try {
-    return readFileSync(markdownPath, "utf-8");
-  } catch (error) {
-    console.warn(`Guide content not found for: ${guideKey}`);
-    return null;
-  }
 }
